Fix build error on OSX
directly use integer version for cp calculation.

fixes https://github.com/official-stockfish/Stockfish/issues/3573

closes https://github.com/official-stockfish/Stockfish/pull/3574

No functional change

+++ src/nnue/evaluate_nnue.cpp
@@ -232,7 +232,7 @@ namespace Stockfish::Eval::NNUE {
 
     buffer[0] = (v < 0 ? '-' : v > 0 ? '+' : ' ');
 
-    int cp = (int)(std::abs(100.0 * double(v) / PawnValueEg));
+    int cp = std::abs(100 * v / PawnValueEg);
 
     if (cp >= 10000)
     {
